fix : 파이썬 노드 데이터 넘기기 이슈 해결 완료
