+++ BE/src/consumption/consumption.service.ts
@@ -6,10 +6,9 @@ export class ConsumptionService {
   private logger = new Logger('ConsumptionService');
 
   async getConsumption(address: string) {
-    const process = spawnSync('python3', ['../AI/consumption-model.py']);
-    const result = JSON.parse(
-      Buffer.from(process.stdout.toJSON().data).toString(),
-    );
+    const process = spawnSync('python3', ['../AI/consumption-main.py']);
+    const str = Buffer.from(process.stdout.toJSON().data).toString();
+    const result = JSON.parse(str.substring(70, str.length));
     return result;
   }
 }
