+++ requirements.txt
@@ -1,0 +1,194 @@
+absl-py==1.0.0
+aiohttp==3.8.1
+aiosignal==1.2.0
+alembic==1.7.6
+appnope==0.1.2
+argon2-cffi==21.3.0
+argon2-cffi-bindings==21.2.0
+asttokens==2.0.5
+astunparse==1.6.3
+async-timeout==4.0.2
+attrs==21.4.0
+autopep8==1.6.0
+backcall==0.2.0
+backports.functools-lru-cache==1.6.4
+beautifulsoup4==4.10.0
+black==22.1.0
+bleach==4.1.0
+blinker==1.4
+brotlipy==0.7.0
+bs4==0.0.1
+cached-property==1.5.2
+cachetools==4.2.4
+certifi==2021.10.8
+cffi==1.15.0
+chardet==4.0.0
+charset-normalizer==2.0.11
+click==8.0.3
+cloudpickle==2.0.0
+cmdstanpy==0.9.5
+convertdate==2.4.0
+cryptography==36.0.1
+cycler==0.11.0
+Cython==0.29.27
+databricks-cli==0.16.4
+dataclasses==0.8
+debugpy==1.5.1
+decorator==5.1.1
+defusedxml==0.7.1
+docker==5.0.3
+entrypoints==0.4
+ephem==4.1.3
+et-xmlfile==1.1.0
+executing==0.8.2
+fbprophet==0.7.1
+Flask==2.0.2
+flatbuffers==2.0
+flit_core==3.6.0
+fonttools==4.29.1
+frozenlist==1.3.0
+future==0.18.2
+gast==0.4.0
+gitdb==4.0.9
+GitPython==3.1.26
+google-auth==1.35.0
+google-auth-oauthlib==0.4.6
+google-pasta==0.2.0
+grpcio==1.42.0
+gunicorn==20.1.0
+h5py==3.6.0
+hijri-converter==2.2.2
+holidays==0.12
+html5lib==1.1
+idna==3.3
+importlib-metadata==4.11.0
+importlib-resources==5.4.0
+ipykernel==6.9.0
+ipython==8.0.1
+ipython-genutils==0.2.0
+itsdangerous==2.0.1
+jedi==0.18.1
+Jinja2==3.0.3
+joblib==1.1.0
+jsonschema==4.4.0
+jupyter-client==7.1.2
+jupyter-core==4.9.1
+jupyterlab-pygments==0.1.2
+keras==2.7.0
+Keras-Preprocessing==1.1.2
+kiwisolver==1.3.2
+korean-lunar-calendar==0.2.1
+lml==0.1.0
+LunarCalendar==0.0.9
+lxml==4.7.1
+Mako==1.1.6
+Markdown==3.3.6
+MarkupSafe==2.0.1
+matplotlib==3.5.1
+matplotlib-inline==0.1.3
+mistune==0.8.4
+mlflow==1.23.1
+multidict==6.0.2
+munkres==1.1.4
+mypy-extensions==0.4.3
+nbclient==0.5.10
+nbconvert==6.4.2
+nbformat==5.1.3
+nest-asyncio==1.5.4
+notebook==6.4.8
+numpy==1.22.2
+oauthlib==3.2.0
+openpyxl==3.0.9
+opt-einsum==3.3.0
+packaging==21.3
+pandas==1.4.1
+pandocfilters==1.5.0
+parso==0.8.3
+pathspec==0.9.0
+patsy==0.5.2
+pexpect==4.8.0
+pickleshare==0.7.5
+Pillow==9.0.1
+pip==22.0.3
+platformdirs==2.4.1
+plotly==5.5.0
+prometheus-client==0.13.1
+prometheus-flask-exporter==0.18.7
+prompt-toolkit==3.0.26
+prophet==1.0.1
+protobuf==3.19.4
+ptyprocess==0.7.0
+pure-eval==0.2.2
+pyasn1==0.4.8
+pyasn1-modules==0.2.7
+pycodestyle==2.8.0
+pycparser==2.21
+pyexcel==0.6.7
+pyexcel-io==0.6.6
+pyexcel-xls==0.7.0
+pyexcel-xlsx==0.6.0
+Pygments==2.11.2
+PyJWT==2.3.0
+PyMeeus==0.5.11
+pyOpenSSL==22.0.0
+pyparsing==3.0.7
+pyrsistent==0.18.1
+PySocks==1.7.1
+pystan==2.19.1.1
+python-dateutil==2.8.2
+pytz==2021.3
+pyu2f==0.1.5
+PyYAML==6.0
+pyzmq==22.3.0
+querystring-parser==1.2.4
+requests==2.27.1
+requests-oauthlib==1.3.1
+rsa==4.8
+scikit-learn==1.0.2
+scipy==1.8.0
+Send2Trash==1.8.0
+setuptools==60.8.2
+setuptools-git==1.2
+six==1.16.0
+sklearn==0.0
+smmap==5.0.0
+soupsieve==2.3.1
+SQLAlchemy==1.4.31
+sqlparse==0.4.2
+stack-data==0.1.4
+statsmodels==0.13.2
+tabulate==0.8.9
+tenacity==8.0.1
+tensorboard==2.6.0
+tensorboard-data-server==0.6.0
+tensorboard-plugin-wit==1.8.1
+tensorflow==2.7.0
+tensorflow-estimator==2.7.0
+termcolor==1.1.0
+terminado==0.13.1
+testpath==0.5.0
+texttable==1.6.4
+threadpoolctl==3.1.0
+toml==0.10.2
+tomli==2.0.1
+torch==1.10.2
+torchvision==0.2.2
+tornado==6.1
+tqdm==4.62.3
+traitlets==5.1.1
+typed-ast==1.5.2
+typing_extensions==4.0.1
+ujson==5.1.0
+unicodedata2==14.0.0
+urllib3==1.26.8
+wcwidth==0.2.5
+webencodings==0.5.1
+websocket-client==1.2.3
+Werkzeug==2.0.3
+wheel==0.37.1
+wrapt==1.13.3
+xgboost==1.5.1
+xlrd==2.0.1
+xlwt==1.3.0
+yarl==1.7.2
+zipp==3.7.0
